docs: document building simulation guidelines
