+++ packages/engine/src/simulation/visualEffectsSystem.ts
@@ -1,5 +1,5 @@
 import type { SimResources } from '../index';
-import type { SimulatedBuilding } from './buildingSimulation';
+import type { SimulatedBuilding } from './buildings/types';
 import type { Citizen } from './citizenBehavior';
 import type { WorkerProfile } from './workers/types';
 import type { VisualIndicator } from './gameplayEvents';
